+++ src/display/displayable_tree.rs
@@ -285,19 +285,35 @@ impl<'a, 's, 't> DisplayableTree<'a, 's, 't> {
             cw.queue_char(style, ' ')?;
             cw.queue_char(style, ' ')?;
         }
-        let label = if pattern_object.subpath {
-            &line.subpath
+        if pattern_object.subpath && line.unlisted == 0 {
+            let name_match = self.tree.options.pattern.pattern
+                .search_string(&line.subpath);
+            let mut path_ms = MatchedString::new(
+                name_match,
+                &line.subpath,
+                &style,
+                &char_match_style,
+            );
+            let name_ms = path_ms.split_on_last('/');
+            cond_bg!(parent_style, self, selected, self.skin.parent);
+            if name_ms.is_some() {
+                path_ms.base_style = &parent_style;
+            }
+            path_ms.queue_on(cw)?;
+            if let Some(name_ms) = name_ms {
+                name_ms.queue_on(cw)?;
+            }
         } else {
-            &line.name
-        };
-        let name_match = self.tree.options.pattern.pattern.search_string(label);
-        let matched_string = MatchedString::new(
-            name_match,
-            label,
-            &style,
-            &char_match_style,
-        );
-        matched_string.queue_on(cw)?;
+            let name_match = self.tree.options.pattern.pattern
+                .search_string(&line.name);
+            let matched_string = MatchedString::new(
+                name_match,
+                &line.name,
+                &style,
+                &char_match_style,
+            );
+            matched_string.queue_on(cw)?;
+        }
         match &line.line_type {
             TreeLineType::Dir => {
                 if line.unlisted > 0 {