+++ src/pattern/pos.rs
@@ -5,3 +5,4 @@ use {
 /// a vector of indexes of the matching characters (not bytes)
 pub type Pos = SmallVec<[usize; 8]>;
 
+