+++ src/skin/style_map.rs
@@ -139,7 +139,7 @@ impl StyleMap {
 StyleMap! {
     default: gray(22), gray(2), [] / gray(20), gray(2), []
     tree: gray(8), None, [] / gray(4), None, []
-    parent: gray(17), None, [] / gray(13), None, []
+    parent: gray(18), None, [] / gray(13), None, []
     file: gray(22), None, [] / gray(15), None, []
     directory: ansi(110), None, [Bold] / ansi(110), None, []
     exe: Some(Cyan), None, []